// GAD-AI-Terminal/libs/bsc/src/contracts.d.ts
export declare const ADDRESSES: {
    WBNB: string;
    PANCAKESWAP_ROUTER_V2: string;
    PANCAKESWAP_FACTORY: string;
    FOUR_MEME: string;
    USDT: string;
};
export declare const PANCAKE_ROUTER_ABI: string[];
export declare const ERC20_ABI: string[];
export declare const PANCAKE_FACTORY_ABI: string[];
export declare const PANCAKE_PAIR_ABI: string[];